Isolate glog include and depend on libglog for #652
