+++ include/osquery/database/results.h
@@ -242,102 +242,6 @@ Status deserializeHistoricalQueryResults(
 Status deserializeHistoricalQueryResultsJSON(const std::string& json,
                                              HistoricalQueryResults& r);
 
-/////////////////////////////////////////////////////////////////////////////
-// ScheduledQueryLogItem
-/////////////////////////////////////////////////////////////////////////////
-
-/**
- * @brief A data structure which represents data to log in the event of an
- * operating system state change
- *
- * When a scheduled query yields new results, we need to log that information
- * to our upstream logging receiver. The data that needs to be logged is the
- * entire DiffResults set as well as some additional metadata.
- */
-struct ScheduledQueryLogItem {
-  /// The data which was changed as a result of the scheduled query
-  DiffResults diffResults;
-
-  /// The name of the scheduled query
-  std::string name;
-
-  /// The identifier (hostname, or uuid) of the host on which the query was
-  /// executed
-  std::string hostIdentifier;
-
-  /// The time that the query was executed, in unix time
-  int unixTime;
-
-  /// The time that the query was executed, in ASCII
-  std::string calendarTime;
-
-  /// equals operator
-  bool operator==(const ScheduledQueryLogItem& comp) const {
-    return (comp.diffResults == diffResults) && (comp.name == name);
-  }
-
-  /// not equals operator
-  bool operator!=(const ScheduledQueryLogItem& comp) const {
-    return !(*this == comp);
-  }
-};
-
-/**
- * @brief Serialize a ScheduledQueryLogItem object into a property tree
- *
- * @param i the ScheduledQueryLogItem to serialize
- * @param tree a reference to a property tree which, if all operations are
- * completed successfully, the contents of ScheduledQueryLogItem will be
- * serialized into
- *
- * @return an instance of osquery::Status, indicating the success or failure
- * of the operation
- */
-Status serializeScheduledQueryLogItem(const ScheduledQueryLogItem& i,
-                                      boost::property_tree::ptree& tree);
-
-/**
- * @brief Serialize a ScheduledQueryLogItem object into a JSON string
- *
- * @param i the ScheduledQueryLogItem to serialize
- * @param json a reference to a string which, if all operations are completed
- * successfully, the contents of ScheduledQueryLogItem will be serialized into
- *
- * @return an instance of osquery::Status, indicating the success or failure
- * of the operation
- */
-Status serializeScheduledQueryLogItemJSON(const ScheduledQueryLogItem& i,
-                                          std::string& json);
-
-/**
- * @brief Serialize a ScheduledQueryLogItem object into a property tree
- * of events, a list of actions.
- *
- * @param item the ScheduledQueryLogItem to serialize
- * @param tree a reference to a property tree which, if all operations are
- * completed successfully, the contents of ScheduledQueryLogItem will be
- * serialized into
- *
- * @return an instance of osquery::Status, indicating the success or failure
- * of the operation
- */
-Status serializeScheduledQueryLogItemAsEvents(
-    const ScheduledQueryLogItem& item, boost::property_tree::ptree& tree);
-
-/**
- * @brief Serialize a ScheduledQueryLogItem object into a JSON string of events,
- * a list of actions.
- *
- * @param i the ScheduledQueryLogItem to serialize
- * @param json a reference to a string which, if all operations are completed
- * successfully, the contents of ScheduledQueryLogItem will be serialized into
- *
- * @return an instance of osquery::Status, indicating the success or failure
- * of the operation
- */
-Status serializeScheduledQueryLogItemAsEventsJSON(
-    const ScheduledQueryLogItem& i, std::string& json);
-
 /**
  * @brief Add a Row to a QueryData if the Row hasn't appeared in the QueryData
  * already
