+++ osquery/config/CMakeLists.txt
@@ -1,7 +1,6 @@
 ADD_OSQUERY_CORE_LIBRARY(osquery_config
   config.cpp
   plugins/filesystem.cpp
-  plugins/glog.cpp
 )
 
 ADD_OSQUERY_TEST(config_tests config_tests.cpp)
