+++ osquery/core/sqlite_util.cpp
@@ -8,11 +8,10 @@
  *
  */
 
+#include <osquery/core.h>
 #include <osquery/database.h>
 #include <osquery/logger.h>
 #include <osquery/tables.h>
-
-#include <osquery/core.h>
 
 #include "osquery/core/sqlite_util.h"
 #include "osquery/core/virtual_table.h"
@@ -36,7 +35,7 @@ QueryData query(const std::string& q, int& error_return) {
 QueryData query(const std::string& q, int& error_return, sqlite3* db) {
   QueryData d;
   char* err = nullptr;
-  sqlite3_exec(db, q.c_str(), core::query_data_callback, &d, &err);
+  sqlite3_exec(db, q.c_str(), query_data_callback, &d, &err);
   if (err != nullptr) {
     LOG(ERROR) << "Error launching query: " << err;
     error_return = 1;
@@ -47,8 +46,6 @@ QueryData query(const std::string& q, int& error_return, sqlite3* db) {
 
   return d;
 }
-
-namespace core {
 
 int query_data_callback(void* argument,
                         int argc,
@@ -65,6 +62,5 @@ int query_data_callback(void* argument,
   }
   (*qData).push_back(r);
   return 0;
-}
 }
 }
