+++ osquery/core/sqlite_util.h
@@ -11,10 +11,8 @@
 #pragma once
 
 namespace osquery {
-namespace core {
 
 // the callback for populating a std::vector<row> set of results. "argument"
 // should be a non-const reference to a std::vector<row>
 int query_data_callback(void *argument, int argc, char *argv[], char *column[]);
-}
 }
