+++ osquery/core/test_util_tests.cpp
@@ -9,15 +9,15 @@
  */
 
 #include <gtest/gtest.h>
-#include <glog/logging.h>
 
 #include <sqlite3.h>
 
 #include <osquery/core.h>
+#include <osquery/logger.h>
+
 #include "osquery/core/test_util.h"
 
 namespace osquery {
-namespace core {
 
 class TestUtilTests : public testing::Test {};
 
@@ -46,7 +46,6 @@ TEST_F(TestUtilTests, test_get_test_db_result_stream) {
     EXPECT_EQ(expected, r.second);
   }
   sqlite3_close(db);
-}
 }
 }
