+++ osquery/core/text_tests.cpp
@@ -8,15 +8,14 @@
  *
  */
 
-#include <osquery/core.h>
-
 #include <gtest/gtest.h>
-#include <glog/logging.h>
+
+#include <osquery/core.h>
+#include <osquery/logger.h>
 
 #include "osquery/core/test_util.h"
 
 namespace osquery {
-namespace core {
 
 class TextTests : public testing::Test {};
 
@@ -24,7 +23,6 @@ TEST_F(TextTests, test_split) {
   for (const auto& i : generateSplitStringTestData()) {
     EXPECT_EQ(split(i.test_string), i.test_vector);
   }
-}
 }
 }
