+++ osquery/database/results_tests.cpp
@@ -12,16 +12,14 @@
 #include <string>
 #include <vector>
 
-#include <glog/logging.h>
 #include <gtest/gtest.h>
 
 #include <osquery/database/results.h>
+#include <osquery/logger.h>
 
 #include "osquery/core/test_util.h"
 
 namespace pt = boost::property_tree;
-
-using namespace osquery::core;
 
 namespace osquery {
