+++ osquery/events/darwin/fsevents.cpp
@@ -10,10 +10,10 @@
 
 #include <boost/numeric/ublas/matrix.hpp>
 
-#include <glog/logging.h>
+#include <osquery/logger.h>
+#include <osquery/tables.h>
 
 #include "osquery/events/darwin/fsevents.h"
-#include <osquery/tables.h>
 
 namespace osquery {
