+++ osquery/logger/plugins/filesystem.cpp
@@ -11,10 +11,9 @@
 #include <exception>
 #include <mutex>
 
-#include <glog/logging.h>
-
 #include <osquery/filesystem.h>
 #include <osquery/flags.h>
+#include <osquery/logger.h>
 #include <osquery/logger/plugin.h>
 
 using osquery::Status;
