+++ osquery/tables/networking/etc_services.cpp
@@ -14,9 +14,8 @@
 #include <boost/algorithm/string/join.hpp>
 #include <boost/algorithm/string/predicate.hpp>
 
-#include <glog/logging.h>
-
 #include <osquery/core.h>
+#include <osquery/logger.h>
 #include <osquery/tables.h>
 #include <osquery/filesystem.h>
