+++ osquery/tables/system/darwin/ca_certs_tests.cpp
@@ -19,8 +19,6 @@
 #include "osquery/core/conversions.h"
 #include "osquery/core/test_util.h"
 
-using namespace osquery::core;
-
 namespace osquery {
 namespace tables {
