+++ osquery/tables/system/darwin/firewall_tests.cpp
@@ -9,15 +9,14 @@
  */
 
 #include <gtest/gtest.h>
-#include <glog/logging.h>
 
 #include <osquery/database.h>
 #include <osquery/filesystem.h>
+#include <osquery/logger.h>
 
 #include "osquery/core/test_util.h"
 #include "osquery/tables/system/darwin/firewall.h"
 
-using namespace osquery::core;
 namespace pt = boost::property_tree;
 
 namespace osquery {
