+++ osquery/tables/system/darwin/homebrew_packages.cpp
@@ -12,11 +12,10 @@
 #include <boost/filesystem/operations.hpp>
 #include <boost/filesystem/path.hpp>
 
-#include <glog/logging.h>
-
 #include <osquery/core.h>
-#include <osquery/tables.h>
 #include <osquery/filesystem.h>
+#include <osquery/logger.h>
+#include <osquery/tables.h>
 
 namespace pt = boost::property_tree;
