+++ osquery/tables/system/darwin/launchd_tests.cpp
@@ -9,14 +9,13 @@
  */
 
 #include <gtest/gtest.h>
-#include <glog/logging.h>
 
 #include <osquery/database.h>
 #include <osquery/filesystem.h>
+#include <osquery/logger.h>
 
 #include "osquery/core/test_util.h"
 
-using namespace osquery::core;
 namespace pt = boost::property_tree;
 
 namespace osquery {
