+++ osquery/tables/system/darwin/nfs_shares.cpp
@@ -6,11 +6,10 @@
 #include <boost/algorithm/string/join.hpp>
 #include <boost/algorithm/string/predicate.hpp>
 
-#include <glog/logging.h>
-
 #include <osquery/core.h>
-#include <osquery/tables.h>
 #include <osquery/filesystem.h>
+#include <osquery/logger.h>
+#include <osquery/tables.h>
 
 namespace osquery {
 namespace tables {
@@ -31,7 +30,6 @@ QueryData parseNfsSharesContent(const std::string& content) {
       index_of_options++;
       if (iter[0] == '/') {
         line_exports.push_back(iter);
-
       } else {
         break;
       }
