+++ osquery/tables/system/darwin/users.mm
@@ -15,11 +15,10 @@
 
 #import <OpenDirectory/OpenDirectory.h>
 
-#include <glog/logging.h>
-
 #include <osquery/core.h>
-#include <osquery/tables.h>
 #include <osquery/filesystem.h>
+#include <osquery/logger.h>
+#include <osquery/tables.h>
 
 namespace osquery {
 namespace tables {
